build: Move idt to a separate directory
